New modual for KernelConnectionDialog
